Improve Roku TV power support
Roku TVs can be turned on and off, but it's helpful for consumers of the
API to be able to determine whether they're talking to a TV and also to
figure out what the current state of the TV is. This patch extracts the
device type (Box, Stick or TV) from the device info and exports that, and
also provides a new power_state method to allow querying of whether the TV
is on or off. It also adds the documented PowerOff command supported on
TVs.